# server/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/nicovideo/download/video:
    post:
      operationId: NicovideoController_downloadVideo
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadVideoRequestDto'
      responses:
        '201':
          description: ''
      tags:
        - Nicovideo
  /api/nicovideo/download/comments:
    post:
      operationId: NicovideoController_downloadComments
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadCommentRequestDto'
      responses:
        '201':
          description: ''
      tags:
        - Nicovideo
info:
  title: danmaku API
  description: Danmaku API documentation
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
    DownloadVideoRequestDto:
      type: object
      properties: {}
    DownloadCommentRequestDto:
      type: object
      properties: {}